+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add `Allow recalls to extend overdue loans` setting. Refs UICIRC-525.
 * Duplicate patron notice template names allowed - case is not considered in validation. Refs UICIRC-528.
 * Circ rules editor does not generate space as expected when adding criteria to multiple criteria rule. Refs UICIRC-488.
+* Add aged to lost triggers to notice policy. Refs UICIRC-515.
 
 ## 4.0.0 (https://github.com/folio-org/ui-circulation/tree/v4.0.0) (2020-10-13)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v3.0.0...v4.0.0)